test: extract shared interop test fixture
The Dna and Vsto test projects carried four verbatim copies of the
Excel fixture (application setup, pid capture, AddWorkbook, the
crash-tolerant teardown). Single inheritance forbids one common base
for all four - they derive from two different fact bases - so the
logic moves to a fixture object fronted by one thin abstract base per
test kind.

New Allors.Excel.Interop.Tests.Shared project (net462, not packable):
InteropExcelFixture owns all the logic; InteropWorksheetTests and
InteropWorkbookTests delegate to it. The per-project classes are now
empty shells whose namespaces keep the per-suite test names in CI
reporting.
