+++ src/Allors.Excel.Interop.Dna.Tests/Allors.Excel.Interop.Dna.Tests.csproj
@@ -10,6 +10,7 @@
   
   <ItemGroup>
     <ProjectReference Include="..\Allors.Excel.Tests\Allors.Excel.Tests.csproj" />
+    <ProjectReference Include="..\Allors.Excel.Interop.Tests.Shared\Allors.Excel.Interop.Tests.Shared.csproj" />
     <ProjectReference Include="..\Allors.Excel.Interop.Dna\Allors.Excel.Interop.Dna.csproj" />
   </ItemGroup>
