+++ src/Allors.Excel.Interop.Vsto.Tests/Allors.Excel.Interop.Vsto.Tests.csproj
@@ -10,6 +10,7 @@
   
   <ItemGroup>
     <ProjectReference Include="..\Allors.Excel.Tests\Allors.Excel.Tests.csproj" />
+    <ProjectReference Include="..\Allors.Excel.Interop.Tests.Shared\Allors.Excel.Interop.Tests.Shared.csproj" />
     <ProjectReference Include="..\Allors.Excel.Interop.Vsto\Allors.Excel.Interop.Vsto.csproj" />
   </ItemGroup>
